Switch to FileContentsFlag for generic oAuth CACert
